+++ src/utils/enum.py
@@ -17,6 +17,7 @@ class Extension(StrEnum):
     CHAMPIONSHIP_DB = 'scch'
     LEGACY_EVENT_DB = 'db'
     ARCHIVE = 'sca'
+    SNAPSHOT = 'scs'
     BACKUP = 'backup'
     SOURCE_DB = 'db'
     TEMPLATE = 'template'